requirements.txt
# ACI — Agent-Computer Interface
# Core (required)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
httpx>=0.25.0
aiofiles>=23.2.0

# Web Bridge (required for web automation)
playwright>=1.41.0

# Desktop Bridge (Windows only — auto-skipped on other platforms)
uiautomation>=2.0.18; sys_platform == "win32"
pywin32>=306; sys_platform == "win32"

# Vision & Memory (required for desktop vision fallback + muscle memory)
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0

# OCR (optional — improves desktop text detection accuracy)
# Uncomment if you want cross-validation OCR:
# rapidocr-onnxruntime>=1.3.0

# PyYAML (for knowledge base loading)
pyyaml>=6.0
